+++ lib/karafka/processing/inline_insights/consumer.rb
@@ -20,10 +20,12 @@ module Karafka
           Tracker.exists?(topic, partition)
         end
 
+        # @return [Hash] alias for `#insights`
         def statistics
           insights
         end
 
+        # @return [Hash] alias for `#insights?`
         def statistics?
           insights?
         end